Loads static and tracks profit in report view

+++ inventory/models.py
@@ -22,7 +22,7 @@ class Ingredient(models.Model):
         ('None', None)
     ]
     unit = models.CharField(max_length=10, choices=MEASUREMENTS)
-    unit_price = models.FloatField()
+    unit_price = models.DecimalField(max_digits=30, decimal_places=2)
 
     def __str__(self):
         return self.name + " quantity: " + str(self.quantity) + ",\n" + "price per " + self.unit + ": $" + str(format(self.unit_price, '.2f')) if (self.unit != 'None') else self.name + " quantity: " + str(self.quantity) + ",\n" + "price per: $" + str(format(self.unit_price, '.2f'))
@@ -32,7 +32,7 @@ class Ingredient(models.Model):
 
 class MenuItem(models.Model):
     title = models.CharField(max_length=30)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=30, decimal_places=2)
 
     def __str__(self):
         return self.title + " costs $" + str(format(self.price, '.2f'))
